refactor(core): move extensions to core
