+++ example/view-based/build.gradle.kts
@@ -39,9 +39,7 @@ android {
 
 dependencies {
 
-    implementation(project(":highlight:core"))
     implementation(project(":highlight:view-based"))
-    implementation(project(":highlight:core-ktx"))
 
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.appcompat)
