+++ example/view-based/src/main/kotlin/com/neoutils/hightlight/example/view/example/TextColorSchemeExample.kt
@@ -3,7 +3,7 @@ package com.neoutils.hightlight.example.view.example
 import android.content.Context
 import android.util.AttributeSet
 import androidx.appcompat.widget.AppCompatTextView
-import com.neoutils.highlight.core.ktx.highlight
+import com.neoutils.highlight.core.extension.highlight
 import com.neoutils.highlight.core.utils.UiColor
 import com.neoutils.highlight.view.extension.toSpannedString
