+++ example/view-based/src/main/kotlin/com/neoutils/hightlight/example/view/example/TextFontSchemeExample.kt
@@ -4,7 +4,7 @@ import android.content.Context
 import android.util.AttributeSet
 import androidx.appcompat.widget.AppCompatTextView
 import androidx.core.content.res.ResourcesCompat
-import com.neoutils.highlight.core.ktx.highlight
+import com.neoutils.highlight.core.extension.highlight
 import com.neoutils.highlight.core.utils.UiFont
 import com.neoutils.highlight.view.extension.toSpannedString
 import com.neoutils.hightlight.example.view.R
