--- highlight/core-ktx/.gitignore
@@ -1,1 +1,0 @@
-/build
