+++ highlight/core/src/main/kotlin/com/neoutils/highlight/core/extension/Highlight.kt
@@ -1,7 +1,7 @@
-package com.neoutils.highlight.core.ktx
+package com.neoutils.highlight.core.extension
 
 import com.neoutils.highlight.core.Highlight
-import com.neoutils.highlight.core.ktx.scope.HighlightScope
+import com.neoutils.highlight.core.scope.HighlightScope
 
 fun highlight(
     scope: HighlightScope.() -> Unit
