+++ highlight/core/src/main/kotlin/com/neoutils/highlight/core/scope/TextStyleScope.kt
@@ -1,4 +1,4 @@
-package com.neoutils.highlight.core.ktx.scope
+package com.neoutils.highlight.core.scope
 
 import com.neoutils.highlight.core.scheme.TextStyleScheme
 import com.neoutils.highlight.core.utils.Match
