+++ highlight/view-based/build.gradle.kts
@@ -31,9 +31,9 @@ android {
 
 dependencies {
 
+    api(project(":highlight:core"))
+
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.appcompat)
     implementation(libs.material)
-
-    implementation(project(":highlight:core"))
 }
